+++ auth_system/urls.py
@@ -7,6 +7,4 @@ urlpatterns = [
    path("logout/", views.logout_view, name='logout'),
    path("register/", views.register, name='register'),
    path("profile/", views.profile, name='profile'),
-   path("novations/", views.novations, name='novations'),
-   path("announs/", views.announs, name='announs'),
 ]
