+++ news/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+   path("novations/", views.novations, name='novations'),
+   path("announs/", views.announs, name='announs'),
+]
